# Make u128 range use widening multiply